engine/cosmetic: move pod event handling code into its own file b/c there's a lot of it
